test(cli): move ignored oxlint config coverage to its own file
